requirements.txt
numpy>=1.24.0,<2.0
pandas>=2.0.3,<3.0
psycopg2-binary==2.9.9
gspread==5.11.3
google-auth-oauthlib==1.1.0
google-auth==2.25.2
python-dotenv==1.0.0
